seperated intake and extension subsystems and also made buttons
