+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,13 +7,13 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "TB2026";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = -1;
-  public static final String GIT_SHA = "UNKNOWN";
-  public static final String GIT_DATE = "UNKNOWN";
-  public static final String GIT_BRANCH = "UNKNOWN";
-  public static final String BUILD_DATE = "2026-01-17 12:24:25 EST";
-  public static final long BUILD_UNIX_TIME = 1768670665129L;
-  public static final int DIRTY = 129;
+  public static final int GIT_REVISION = 3;
+  public static final String GIT_SHA = "69ef7e5e6817c4bd534863bdd82fb57a8c0d6bd6";
+  public static final String GIT_DATE = "2026-01-17 16:43:45 EST";
+  public static final String GIT_BRANCH = "Intake";
+  public static final String BUILD_DATE = "2026-01-20 19:32:47 EST";
+  public static final long BUILD_UNIX_TIME = 1768955567855L;
+  public static final int DIRTY = 1;
 
   private BuildConstants(){}
 }
