+++ src/main/java/frc/robot/subsystems/extension/ExtensionIOSim.java
@@ -1,0 +1,8 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.subsystems.extension;
+
+/** Add your docs here. */
+public class ExtensionIOSim implements ExtensionIO{}
