+++ src/main/java/frc/robot/subsystems/intake/IntakeIO.java
@@ -19,20 +19,9 @@ public interface IntakeIO {
     public boolean intakeEncoderConnected = false;
     public double intakeSpeed = 0.0;
 
-    public double extensionAngle = 0.0;
-    public double extensionSetpoint = 0.0;
-    public double extensionStatorCurrent = 0.0;
-    public double extensionSupplyCurrent = 0.0;
-    public double extensionVoltage = 0.0;
-    public double extensionMotorTemp = 0.0;
-    public boolean extensionMotorConnected = false;
-    public boolean extensionEncoderConnected = false;
-    public double extensionSpeed = 0.0;
   }
   
   public default void updateInputs(IntakeIOInputs inputs) {}
-
-  public default void setExtensionSetpoint(double position) {}
 
   public default void setOutput(double voltage) {}
 }
